refactor: unify did document services
