+++ packages/core/src/agent/TransportService.ts
@@ -1,5 +1,6 @@
-import type { DidDoc, IndyAgentService } from '../modules/connections/models'
+import type { DidDoc } from '../modules/connections/models'
 import type { ConnectionRecord } from '../modules/connections/repository'
+import type { IndyAgentService } from '../modules/dids/domain/service'
 import type { EncryptedMessage } from '../types'
 import type { AgentMessage } from './AgentMessage'
 import type { EnvelopeKeys } from './EnvelopeService'
@@ -7,7 +8,8 @@ import type { EnvelopeKeys } from './EnvelopeService'
 import { Lifecycle, scoped } from 'tsyringe'
 
 import { DID_COMM_TRANSPORT_QUEUE } from '../constants'
-import { ConnectionRole, DidCommService } from '../modules/connections/models'
+import { ConnectionRole } from '../modules/connections/models'
+import { DidCommService } from '../modules/dids/domain/service'
 
 @scoped(Lifecycle.ContainerScoped)
 export class TransportService {
