+++ packages/core/src/modules/connections/models/did/index.ts
@@ -1,4 +1,3 @@
 export * from './DidDoc'
-export * from './service'
 export * from './publicKey'
 export * from './authentication'
