--- packages/core/src/modules/connections/models/did/service/IndyAgentService.ts
@@ -1,33 +1,0 @@
-import { ArrayNotEmpty, IsOptional, IsString } from 'class-validator'
-
-import { Service } from './Service'
-
-export class IndyAgentService extends Service {
-  public constructor(options: {
-    id: string
-    serviceEndpoint: string
-    recipientKeys: string[]
-    routingKeys?: string[]
-    priority?: number
-  }) {
-    super({ ...options, type: IndyAgentService.type })
-
-    if (options) {
-      this.recipientKeys = options.recipientKeys
-      this.routingKeys = options.routingKeys
-      if (options.priority) this.priority = options.priority
-    }
-  }
-
-  public static type = 'IndyAgent'
-
-  @ArrayNotEmpty()
-  @IsString({ each: true })
-  public recipientKeys!: string[]
-
-  @IsString({ each: true })
-  @IsOptional()
-  public routingKeys?: string[]
-
-  public priority = 0
-}
