--- packages/core/src/modules/connections/models/did/service/index.ts
@@ -1,39 +1,0 @@
-import type { ClassConstructor } from 'class-transformer'
-
-import { Transform, plainToInstance } from 'class-transformer'
-
-import { DidCommService } from './DidCommService'
-import { IndyAgentService } from './IndyAgentService'
-import { Service } from './Service'
-
-export const serviceTypes: { [key: string]: unknown | undefined } = {
-  [IndyAgentService.type]: IndyAgentService,
-  [DidCommService.type]: DidCommService,
-}
-
-/**
- * Decorator that transforms service json to corresponding class instances. See {@link serviceTypes}
- *
- * @example
- * class Example {
- *   ServiceTransformer()
- *   private service: Service
- * }
- */
-export function ServiceTransformer() {
-  return Transform(
-    ({ value }: { value: { type: string }[] }) => {
-      return value.map((serviceJson) => {
-        const serviceClass = (serviceTypes[serviceJson.type] ?? Service) as ClassConstructor<Service>
-        const service = plainToInstance<Service, unknown>(serviceClass, serviceJson)
-
-        return service
-      })
-    },
-    {
-      toClassOnly: true,
-    }
-  )
-}
-
-export { IndyAgentService, DidCommService, Service }
